change directory structure of server file
Previously all files were in server.js. This makes for a huge file, which could be split up into different components:

* models: for models
* routes: for different routes, seperated by each router
